refactor(pipeline): split pipeline execution into dedicated modules
Extract core pipeline logic from _pipeline_internals.py into three new modules:
- _pipeline_streams.py for stream coordination
- _pipeline_wait.py for pipeline waiting and fail-fast handling
- _process_lifecycle.py for process lifecycle management

This improves code organization and separation of concerns by isolating stream management, process lifecycle, and pipeline wait logic.
