+++ cuprum/_pipeline_streams.py
@@ -1,0 +1,204 @@
+"""Stream coordination for pipeline execution."""
+
+from __future__ import annotations
+
+import asyncio
+import dataclasses as dc
+import sys
+import typing as typ
+
+from cuprum._streams import _consume_stream, _pump_stream, _StreamConfig
+
+if typ.TYPE_CHECKING:
+    from cuprum._pipeline_internals import _StageObservation
+    from cuprum.sh import ExecutionContext
+
+
+@dc.dataclass(frozen=True, slots=True)
+class _PipelineRunConfig:
+    ctx: ExecutionContext
+    capture: bool
+    echo: bool
+    stdout_sink: typ.IO[str]
+    stderr_sink: typ.IO[str]
+
+    @property
+    def capture_or_echo(self) -> bool:
+        return self.capture or self.echo
+
+    @property
+    def stream_config(self) -> _StreamConfig:
+        return _StreamConfig(
+            capture_output=self.capture,
+            echo_output=self.echo,
+            sink=self.stdout_sink,
+            encoding=self.ctx.encoding,
+            errors=self.ctx.errors,
+        )
+
+
+def _prepare_pipeline_config(
+    *,
+    capture: bool,
+    echo: bool,
+    context: ExecutionContext | None,
+) -> _PipelineRunConfig:
+    """Normalise runtime options for pipeline execution."""
+    from cuprum._pipeline_internals import _sh_module
+
+    sh = _sh_module()
+    ctx = context or sh.ExecutionContext()
+    stdout_sink = ctx.stdout_sink if ctx.stdout_sink is not None else sys.stdout
+    stderr_sink = ctx.stderr_sink if ctx.stderr_sink is not None else sys.stderr
+    return _PipelineRunConfig(
+        ctx=ctx,
+        capture=capture,
+        echo=echo,
+        stdout_sink=stdout_sink,
+        stderr_sink=stderr_sink,
+    )
+
+
+@dc.dataclass(frozen=True, slots=True)
+class _StageStreamConfig:
+    """Stream file descriptor configuration for a pipeline stage."""
+
+    stdin: int
+    stdout: int
+    stderr: int
+
+
+def _get_stage_stream_fds(
+    idx: int,
+    last_idx: int,
+    *,
+    capture_or_echo: bool,
+) -> _StageStreamConfig:
+    """Determine stream file descriptors for a pipeline stage.
+
+    First stage reads from DEVNULL; intermediate stages use pipes for stdin.
+    stdout is piped for intermediate stages or when capturing. stderr is piped
+    only when capturing or echoing.
+    """
+    stdin = asyncio.subprocess.DEVNULL if idx == 0 else asyncio.subprocess.PIPE
+    stdout = (
+        asyncio.subprocess.PIPE
+        if idx != last_idx or capture_or_echo
+        else asyncio.subprocess.DEVNULL
+    )
+    stderr = asyncio.subprocess.PIPE if capture_or_echo else asyncio.subprocess.DEVNULL
+    return _StageStreamConfig(stdin=stdin, stdout=stdout, stderr=stderr)
+
+
+def _create_stage_capture_tasks(
+    process: asyncio.subprocess.Process,
+    config: _PipelineRunConfig,
+    *,
+    is_last_stage: bool,
+    observation: _StageObservation,
+) -> tuple[asyncio.Task[str | None] | None, asyncio.Task[str | None] | None]:
+    """Create stderr and stdout capture tasks for a pipeline stage.
+
+    Returns (stderr_task, stdout_task). stderr is captured for all stages when
+    capture_or_echo is enabled. stdout is only captured for the final stage.
+    """
+    stderr_task: asyncio.Task[str | None] | None = None
+    stdout_task: asyncio.Task[str | None] | None = None
+
+    if not config.capture_or_echo:
+        return stderr_task, stdout_task
+
+    stderr_on_line: typ.Callable[[str], None] | None = None
+    if observation.hooks.observe_hooks:
+
+        def stderr_on_line(line: str) -> None:
+            from cuprum._pipeline_internals import _EventDetails
+
+            observation.emit(
+                "stderr",
+                _EventDetails(pid=process.pid, line=line),
+            )
+
+    stderr_task = asyncio.create_task(
+        _consume_stream(
+            process.stderr,
+            dc.replace(config.stream_config, sink=config.stderr_sink),
+            on_line=stderr_on_line,
+        ),
+    )
+
+    if not is_last_stage:
+        return stderr_task, stdout_task
+
+    stdout_on_line: typ.Callable[[str], None] | None = None
+    if observation.hooks.observe_hooks:
+
+        def stdout_on_line(line: str) -> None:
+            from cuprum._pipeline_internals import _EventDetails
+
+            observation.emit(
+                "stdout",
+                _EventDetails(pid=process.pid, line=line),
+            )
+
+    stdout_task = asyncio.create_task(
+        _consume_stream(
+            process.stdout,
+            config.stream_config,
+            on_line=stdout_on_line,
+        ),
+    )
+
+    return stderr_task, stdout_task
+
+
+def _create_pipe_tasks(
+    processes: list[asyncio.subprocess.Process],
+) -> list[asyncio.Task[None]]:
+    """Create streaming tasks between adjacent pipeline stages."""
+    return [
+        asyncio.create_task(
+            _pump_stream(
+                processes[idx].stdout,
+                processes[idx + 1].stdin,
+            ),
+        )
+        for idx in range(len(processes) - 1)
+    ]
+
+
+def _flatten_stream_tasks(
+    stderr_tasks: list[asyncio.Task[str | None] | None],
+    stdout_task: asyncio.Task[str | None] | None,
+) -> list[asyncio.Task[str | None]]:
+    """Collect all running stream consumer tasks for cancellation cleanup."""
+    tasks = [task for task in stderr_tasks if task is not None]
+    if stdout_task is not None:
+        tasks.append(stdout_task)
+    return tasks
+
+
+async def _collect_pipe_results(
+    pipe_tasks: list[asyncio.Task[None]],
+) -> list[object]:
+    """Collect pipe task results, capturing exceptions rather than raising them.
+
+    Uses return_exceptions=True to gather all results including any exceptions
+    that occurred during pipe streaming between pipeline stages.
+    """
+    return list(await asyncio.gather(*pipe_tasks, return_exceptions=True))
+
+
+def _surface_unexpected_pipe_failures(pipe_results: list[object]) -> None:
+    """Raise non-BrokenPipe exceptions from pipe results.
+
+    BrokenPipeError and ConnectionResetError are expected when downstream
+    processes terminate early (e.g., head) and should not fail the pipeline.
+    Other exceptions indicate genuine failures and must be surfaced.
+    """
+    for result in pipe_results:
+        if isinstance(result, Exception) and not isinstance(
+            result,
+            (BrokenPipeError, ConnectionResetError),
+        ):
+            raise result
